+++ edit-post/store/test/selectors.js
@@ -1,4 +1,9 @@
 /**
+ * WordPress dependencies
+ */
+import deprecated from '@wordpress/deprecated';
+
+/**
  * Internal dependencies
  */
 import {
@@ -15,9 +20,17 @@ import {
 	hasMetaBoxes,
 	isSavingMetaBoxes,
 	getMetaBox,
+	getActiveMetaBoxLocations,
+	isMetaBoxLocationActive,
 } from '../selectors';
 
+jest.mock( '@wordpress/deprecated', () => jest.fn() );
+
 describe( 'selectors', () => {
+	beforeEach( () => {
+		deprecated.mockClear();
+	} );
+
 	describe( 'getEditorMode', () => {
 		it( 'should return the selected editor mode', () => {
 			const state = {
@@ -283,14 +296,7 @@ describe( 'selectors', () => {
 	describe( 'hasMetaBoxes', () => {
 		it( 'should return true if there are active meta boxes', () => {
 			const state = {
-				metaBoxes: {
-					normal: {
-						isActive: false,
-					},
-					side: {
-						isActive: true,
-					},
-				},
+				activeMetaBoxLocations: [ 'side' ],
 			};
 
 			expect( hasMetaBoxes( state ) ).toBe( true );
@@ -298,14 +304,7 @@ describe( 'selectors', () => {
 
 		it( 'should return false if there are no active meta boxes', () => {
 			const state = {
-				metaBoxes: {
-					normal: {
-						isActive: false,
-					},
-					side: {
-						isActive: false,
-					},
-				},
+				activeMetaBoxLocations: [],
 			};
 
 			expect( hasMetaBoxes( state ) ).toBe( false );
@@ -333,19 +332,18 @@ describe( 'selectors', () => {
 	describe( 'getMetaBoxes', () => {
 		it( 'should return the state of all meta boxes', () => {
 			const state = {
-				metaBoxes: {
-					normal: {
-						isActive: true,
-					},
-					side: {
-						isActive: true,
-					},
-				},
+				activeMetaBoxLocations: [ 'normal', 'side' ],
 			};
 
-			expect( getMetaBoxes( state ) ).toEqual( {
+			const result = getMetaBoxes( state );
+
+			expect( deprecated ).toHaveBeenCalled();
+			expect( result ).toEqual( {
 				normal: {
 					isActive: true,
+				},
+				advanced: {
+					isActive: false,
 				},
 				side: {
 					isActive: true,
@@ -357,19 +355,49 @@ describe( 'selectors', () => {
 	describe( 'getMetaBox', () => {
 		it( 'should return the state of selected meta box', () => {
 			const state = {
-				metaBoxes: {
-					normal: {
-						isActive: false,
-					},
-					side: {
-						isActive: true,
-					},
-				},
+				activeMetaBoxLocations: [ 'side' ],
 			};
 
-			expect( getMetaBox( state, 'side' ) ).toEqual( {
+			const result = getMetaBox( state, 'side' );
+
+			expect( deprecated ).toHaveBeenCalled();
+			expect( result ).toEqual( {
 				isActive: true,
 			} );
+		} );
+	} );
+
+	describe( 'getActiveMetaBoxLocations', () => {
+		it( 'should return the active meta boxes', () => {
+			const state = {
+				activeMetaBoxLocations: [ 'side' ],
+			};
+
+			const result = getActiveMetaBoxLocations( state, 'side' );
+
+			expect( result ).toEqual( [ 'side' ] );
+		} );
+	} );
+
+	describe( 'isMetaBoxLocationActive', () => {
+		it( 'should return false if not active', () => {
+			const state = {
+				activeMetaBoxLocations: [],
+			};
+
+			const result = isMetaBoxLocationActive( state, 'side' );
+
+			expect( result ).toBe( false );
+		} );
+
+		it( 'should return true if active', () => {
+			const state = {
+				activeMetaBoxLocations: [ 'side' ],
+			};
+
+			const result = isMetaBoxLocationActive( state, 'side' );
+
+			expect( result ).toBe( true );
 		} );
 	} );
 } );