+++ packages/block-library/src/gallery/edit.js
@@ -26,6 +26,7 @@ import {
 	InspectorControls,
 	mediaUpload,
 } from '@wordpress/editor';
+import { create } from '@wordpress/rich-text';
 
 /**
  * Internal dependencies
@@ -43,6 +44,19 @@ const ALLOWED_MEDIA_TYPES = [ 'image' ];
 export function defaultColumnsNumber( attributes ) {
 	return Math.min( 3, attributes.images.length );
 }
+
+export const pickRelevantMediaFiles = ( image ) => {
+	let { caption } = image;
+
+	if ( typeof caption !== 'object' ) {
+		caption = create( { html: caption } );
+	}
+
+	return {
+		...pick( image, [ 'alt', 'id', 'link', 'url' ] ),
+		caption,
+	};
+};
 
 class GalleryEdit extends Component {
 	constructor() {
@@ -88,8 +102,8 @@ class GalleryEdit extends Component {
 	onSelectImages( images ) {
 		const idList = images.map( ( item ) => item.id );
 		this.props.setAttributes( {
-			ids: idList,
-			images: images.map( ( image ) => pick( image, [ 'alt', 'caption', 'id', 'link', 'url' ] ) ),
+      ids: idList,
+			images: images.map( ( image ) => pickRelevantMediaFiles( image ) ),
 		} );
 	}
 
@@ -137,8 +151,9 @@ class GalleryEdit extends Component {
 			allowedTypes: ALLOWED_MEDIA_TYPES,
 			filesList: files,
 			onFileChange: ( images ) => {
+				const imagesNormalized = images.map( ( image ) => pickRelevantMediaFiles( image ) );
 				setAttributes( {
-					images: currentImages.concat( images ),
+					images: currentImages.concat( imagesNormalized ),
 				} );
 			},
 			onError: noticeOperations.createErrorNotice,