+++ packages/block-library/src/image/edit.js
@@ -2,7 +2,6 @@
  * External dependencies
  */
 import classnames from 'classnames';
-import ResizableBox from 're-resizable';
 import {
 	get,
 	isEmpty,
@@ -22,6 +21,7 @@ import {
 	ButtonGroup,
 	IconButton,
 	PanelBody,
+	ResizableBox,
 	SelectControl,
 	TextControl,
 	TextareaControl,
@@ -40,6 +40,7 @@ import {
 } from '@wordpress/editor';
 import { withViewportMatch } from '@wordpress/viewport';
 import { compose } from '@wordpress/compose';
+import { create } from '@wordpress/rich-text';
 
 /**
  * Internal dependencies
@@ -55,6 +56,19 @@ const LINK_DESTINATION_MEDIA = 'media';
 const LINK_DESTINATION_ATTACHMENT = 'attachment';
 const LINK_DESTINATION_CUSTOM = 'custom';
 const ALLOWED_MEDIA_TYPES = [ 'image' ];
+
+export const pickRelevantMediaFiles = ( image ) => {
+	let { caption } = image;
+
+	if ( typeof caption !== 'object' ) {
+		caption = create( { html: caption } );
+	}
+
+	return {
+		...pick( image, [ 'alt', 'id', 'link', 'url' ] ),
+		caption,
+	};
+};
 
 class ImageEdit extends Component {
 	constructor() {
@@ -87,7 +101,7 @@ class ImageEdit extends Component {
 				mediaUpload( {
 					filesList: [ file ],
 					onFileChange: ( [ image ] ) => {
-						setAttributes( { ...image } );
+						setAttributes( pickRelevantMediaFiles( image ) );
 					},
 					allowedTypes: ALLOWED_MEDIA_TYPES,
 				} );
@@ -116,12 +130,13 @@ class ImageEdit extends Component {
 				url: undefined,
 				alt: undefined,
 				id: undefined,
-				caption: undefined,
+				caption: create(),
 			} );
 			return;
 		}
+
 		this.props.setAttributes( {
-			...pick( media, [ 'alt', 'id', 'caption', 'url' ] ),
+			...pickRelevantMediaFiles( media ),
 			width: undefined,
 			height: undefined,
 		} );
@@ -435,7 +450,6 @@ class ImageEdit extends Component {
 								<Fragment>
 									{ getInspectorControls( imageWidth, imageHeight ) }
 									<ResizableBox
-										className="block-library-image__resizer"
 										size={
 											width && height ? {
 												width,
@@ -447,11 +461,6 @@ class ImageEdit extends Component {
 										minHeight={ minHeight }
 										maxHeight={ maxWidth / ratio }
 										lockAspectRatio
-										handleClasses={ {
-											right: 'block-library-image__resize-handler-right',
-											bottom: 'block-library-image__resize-handler-bottom',
-											left: 'block-library-image__resize-handler-left',
-										} }
 										enable={ {
 											top: false,
 											right: showRightHandle,
@@ -479,7 +488,7 @@ class ImageEdit extends Component {
 						<RichText
 							tagName="figcaption"
 							placeholder={ __( 'Write caption…' ) }
-							value={ caption || [] }
+							value={ caption }
 							unstableOnFocus={ this.onFocusCaption }
 							onChange={ ( value ) => setAttributes( { caption: value } ) }
 							isSelected={ this.state.captionFocused }