+++ packages/editor/src/store/defaults.js
@@ -81,27 +81,28 @@ export const EDITOR_SETTINGS_DEFAULTS = {
 	fontSizes: [
 		{
 			name: __( 'Small' ),
-			shortName: __( 'S' ),
-			size: 14,
+			size: 13,
 			slug: 'small',
 		},
 		{
-			name: __( 'Regular' ),
-			shortName: __( 'M' ),
+			name: __( 'Normal' ),
 			size: 16,
-			slug: 'regular',
+			slug: 'normal',
+		},
+		{
+			name: __( 'Medium' ),
+			size: 20,
+			slug: 'medium',
 		},
 		{
 			name: __( 'Large' ),
-			shortName: __( 'L' ),
 			size: 36,
 			slug: 'large',
 		},
 		{
-			name: __( 'Larger' ),
-			shortName: __( 'XL' ),
+			name: __( 'Huge' ),
 			size: 48,
-			slug: 'larger',
+			slug: 'huge',
 		},
 	],