+++ packages/editor/CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Deprecations
 
 - The `checkTemplateValidity` action has been deprecated. Validity is verified automatically upon block reset.
+- The `UnsavedChangesWarning` component `forceIsDirty` prop has been deprecated.
 
 ## 3.0.0 (2018-09-05)